requirements.txt
python-dotenv
requests
pandas
numpy
websocket-client
python-telegram-bot==20.7
